+++ image.hpp
@@ -16,20 +16,20 @@ const ushort incrementS = 4;
 
 class Image {
  
-    vector<vector<int> > data;
+    vector<vector<double> > data;
     int clacc;
-    vector<vector<int> > integral;
+    vector<vector<double> > integral;
     vector<Feature> featureVector;
 
     public:
         Image(string fname, short c);
-        Image(vector<vector<int> > &data, int c);
+        Image(vector<vector<double> > &data, int c);
         void calIntegral();
         void calFeatureVector();
         void initialize();
-        vector<vector<int> > &getImageData();
+        vector<vector<double> > &getImageData();
         int getImageClass();
-        vector<vector<int> > &getImageIntegral();
+        vector<vector<double> > &getImageIntegral();
         vector<Feature> &getFeatureVector();
         void calFeatureByLines(int start, int end, vector<pair<int, int> > &couples);
         int sizeOfFeatures();
